+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fix:
   - Display primary key tip only for array stores.
   - Fix store compilation when using multi stores registration.
+  - A store cannot reference itself.
 
 ## 1.10.0
 *2019-07-01*